+++ zones_test.go
@@ -1,11 +1,12 @@
 package egoscale
 
 import (
+	"context"
 	"testing"
 	"time"
 )
 
-func TestListZones(t *testing.T) {
+func TestListZonesAPIName(t *testing.T) {
 	req := &ListZones{}
 	if req.APIName() != "listZones" {
 		t.Errorf("API call doesn't match")
@@ -13,7 +14,7 @@ func TestListZones(t *testing.T) {
 	_ = req.response().(*ListZonesResponse)
 }
 
-func TestListZone(t *testing.T) {
+func TestListZones(t *testing.T) {
 	ts := newServer(response{200, `
 {"listzonesresponse": {
 	"count": 4,
@@ -83,7 +84,173 @@ func TestListZone(t *testing.T) {
 	}
 }
 
-func TestListZoneTwoPages(t *testing.T) {
+func TestListZonesPaginate(t *testing.T) {
+	ts := newServer(response{200, `
+{"listzonesresponse": {
+	"count": 4,
+	"zone": [
+		{
+			"allocationstate": "Enabled",
+			"dhcpprovider": "VirtualRouter",
+			"id": "1747ef5e-5451-41fd-9f1a-58913bae9702",
+			"localstorageenabled": true,
+			"name": "ch-gva-2",
+			"networktype": "Basic",
+			"securitygroupsenabled": true,
+			"tags": [],
+			"zonetoken": "f9a2983b-42e5-3b12-ae74-0b1f54cd6204"
+		},
+		{
+			"allocationstate": "Enabled",
+			"dhcpprovider": "VirtualRouter",
+			"id": "381d0a95-ed4a-4ad9-b41c-b97073c1a433",
+			"localstorageenabled": true,
+			"name": "ch-dk-2",
+			"networktype": "Basic",
+			"securitygroupsenabled": true,
+			"tags": [],
+			"zonetoken": "23a24359-121a-38af-a938-e225c97c397b"
+		},
+		{
+			"allocationstate": "Enabled",
+			"dhcpprovider": "VirtualRouter",
+			"id": "b0fcd72f-47ad-4779-a64f-fe4de007ec72",
+			"localstorageenabled": true,
+			"name": "at-vie-1",
+			"networktype": "Basic",
+			"securitygroupsenabled": true,
+			"tags": [],
+			"zonetoken": "a2a8345d-7daa-3316-8d90-5b8e49706764"
+		},
+		{
+			"allocationstate": "Enabled",
+			"dhcpprovider": "VirtualRouter",
+			"id": "de88c980-78f6-467c-a431-71bcc88e437f",
+			"localstorageenabled": true,
+			"name": "de-fra-1",
+			"networktype": "Basic",
+			"securitygroupsenabled": true,
+			"tags": [],
+			"zonetoken": "c4bdb9f2-c28d-36a3-bbc5-f91fc69527e6"
+		}
+	]
+}}`})
+	defer ts.Close()
+
+	cs := NewClient(ts.URL, "KEY", "SECRET")
+
+	zone := new(Zone)
+	req, _ := zone.ListRequest()
+
+	counter := 0
+	cs.Paginate(req, func(i interface{}, e error) bool {
+		if e != nil {
+			t.Error(e)
+			return false
+		}
+		z := i.(Zone)
+		if z.ID == "" {
+			t.Errorf("Zone ID not set")
+		}
+		counter++
+		return true
+	})
+
+	if counter != 4 {
+		t.Errorf("Four zones were expected, got %d", counter)
+	}
+}
+
+func TestListZonesAsync(t *testing.T) {
+	ts := newServer(response{200, `
+{"listzonesresponse": {
+	"count": 4,
+	"zone": [
+		{
+			"allocationstate": "Enabled",
+			"dhcpprovider": "VirtualRouter",
+			"id": "1747ef5e-5451-41fd-9f1a-58913bae9702",
+			"localstorageenabled": true,
+			"name": "ch-gva-2",
+			"networktype": "Basic",
+			"securitygroupsenabled": true,
+			"tags": [],
+			"zonetoken": "f9a2983b-42e5-3b12-ae74-0b1f54cd6204"
+		},
+		{
+			"allocationstate": "Enabled",
+			"dhcpprovider": "VirtualRouter",
+			"id": "381d0a95-ed4a-4ad9-b41c-b97073c1a433",
+			"localstorageenabled": true,
+			"name": "ch-dk-2",
+			"networktype": "Basic",
+			"securitygroupsenabled": true,
+			"tags": [],
+			"zonetoken": "23a24359-121a-38af-a938-e225c97c397b"
+		},
+		{
+			"allocationstate": "Enabled",
+			"dhcpprovider": "VirtualRouter",
+			"id": "b0fcd72f-47ad-4779-a64f-fe4de007ec72",
+			"localstorageenabled": true,
+			"name": "at-vie-1",
+			"networktype": "Basic",
+			"securitygroupsenabled": true,
+			"tags": [],
+			"zonetoken": "a2a8345d-7daa-3316-8d90-5b8e49706764"
+		},
+		{
+			"allocationstate": "Enabled",
+			"dhcpprovider": "VirtualRouter",
+			"id": "de88c980-78f6-467c-a431-71bcc88e437f",
+			"localstorageenabled": true,
+			"name": "de-fra-1",
+			"networktype": "Basic",
+			"securitygroupsenabled": true,
+			"tags": [],
+			"zonetoken": "c4bdb9f2-c28d-36a3-bbc5-f91fc69527e6"
+		}
+	]
+}}`})
+	defer ts.Close()
+
+	cs := NewClient(ts.URL, "KEY", "SECRET")
+
+	zone := new(Zone)
+
+	outChan, errChan := cs.AsyncListWithContext(context.TODO(), zone)
+
+	counter := 0
+	for {
+		select {
+		case z, ok := <-outChan:
+			if ok {
+				zone := z.(Zone)
+				if zone.ID == "" {
+					t.Errorf("Zone has no ID")
+				}
+				counter++
+			} else {
+				outChan = nil
+			}
+		case e, ok := <-errChan:
+			if ok {
+				t.Error(e)
+			}
+			errChan = nil
+		}
+
+		if outChan == nil && errChan == nil {
+			break
+		}
+	}
+
+	if counter != 4 {
+		t.Errorf("Four zones were expected, got %d", counter)
+	}
+}
+
+func TestListZonesTwoPages(t *testing.T) {
 	ts := newServer(response{200, `
 {"listzonesresponse": {
 	"count": 4,
@@ -158,7 +325,7 @@ func TestListZoneTwoPages(t *testing.T) {
 	}
 }
 
-func TestListZoneError(t *testing.T) {
+func TestListZonesError(t *testing.T) {
 	ts := newServer(response{200, `
 {"listzonesresponse": {
 	"count": 4,
@@ -205,7 +372,7 @@ func TestListZoneError(t *testing.T) {
 	}
 }
 
-func TestListZoneTimeout(t *testing.T) {
+func TestListZonesTimeout(t *testing.T) {
 	ts := newSleepyServer(time.Second, 200, `
 {"listzonesresponse": {
 	"count": 4